# services/scenario-manager/docs/swagger.yaml
basePath: /
definitions:
...
  entities.Test:
    properties:
      cmd:
        type: string
      name:
        type: string
      parameters:
        items:
          type: string
        type: array
      script:
        type: string
      when_to_run:
        type: string
      where_to_run:
        type: string
    type: object
  entities.TestConfig:
    properties:
      name:
        type: string
      tests:
        items:
          $ref: '#/definitions/entities.Test'
        type: array
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Restful APIs for Merak Cloud Emulator
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Merak Rest API
  version: "2.0"
paths:
...
  /api/test-config:
    get:
      consumes:
      - application/json
      description: Get all test-config
      responses:
        "200":
          description: array of test-config with success message
          schema:
            items:
              $ref: '#/definitions/entities.TestConfig'
            type: array
        "404":
          description: null test-config data with error message
      summary: Get all test-config from database
      tags:
      - test-config
    post:
      consumes:
      - application/json
      description: Create a test-config
      parameters:
      - description: TestConfig
        in: body
        name: test_config
        required: true
        schema:
          $ref: '#/definitions/entities.TestConfig'
      responses:
        "200":
          description: Compute data with success message
          schema:
            $ref: '#/definitions/entities.TestConfig'
        "500":
          description: Compute null with failure message
      summary: Insert a test-config to database
      tags:
      - test-config
  /api/test-config/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a test-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: test-config data with success message
          schema:
            $ref: '#/definitions/entities.TestConfig'
        "404":
          description: test-config data with null and error message
      summary: Delete a test-config from database
      tags:
      - test-config
    get:
      consumes:
      - application/json
      description: Get a test-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: test-config data with success message
          schema:
            $ref: '#/definitions/entities.TestConfig'
        "404":
          description: test-config data with null and error message
      summary: Get a test-config from database
      tags:
      - test-config
    put:
      consumes:
      - application/json
      description: Update a test-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      - description: TestConfig
        in: body
        name: compute_config
        required: true
        schema:
          type: string
      responses:
        "200":
          description: compute_config data with success message
          schema:
            $ref: '#/definitions/entities.TestConfig'
        "500":
          description: compute_config null with failure message
      summary: Update a test-config to database
      tags:
      - test-config
...
schemes:
- http
swagger: "2.0"